+++ project/server.py
@@ -1,0 +1,5 @@
+from flask import Flask
+from dash import Dash
+
+server = Flask('myproject')
+app = Dash(server=server)
