+++ project/wsgi.py
@@ -1,0 +1,1 @@
+from app import server as application
