+++ run.py
@@ -1,0 +1,4 @@
+from project.app import app
+
+if __name__ == '__main__':
+    app.run_server(debug=True)
